Define PacketHeader in the main module, not raw
